move state module to bokeh.core
