+++ bokeh/io.py
@@ -24,13 +24,13 @@ import io, itertools, os, warnings
 
 # Bokeh imports
 from . import browserlib
+from .core.state import State
 from .document import Document
 from .embed import notebook_div, standalone_html_page_for_models, autoload_server
 from .models import Component
 from .models.plots import GridPlot
 from .models.widgets.layouts import HBox, VBox, VBoxForm
 from .model import _ModelInDocument
-from .state import State
 from .util.notebook import load_notebook, publish_display_data
 from .util.string import decode_utf8
 from .client import DEFAULT_SESSION_ID, push_session, show_session
